Fix conflicts. Some minor refactoring
