+++ app/src/main/java/org/mozilla/reference/browser/topsites/ui/TopSitesView.kt
@@ -44,7 +44,6 @@ class TopSitesView @JvmOverloads constructor(
     }
 
     fun initialize(
-        context: Context,
         historyUseCases: HistoryUseCases,
         browserIcons: BrowserIcons
     ) {
